feat: integrate the cheating end point to mark the grade as cheated

+++ lib/Core/data/models/enrolled_course_model.dart
@@ -4,7 +4,7 @@ class EnrolledCourseModel {
   final CourseModel courseModel;
   final String id;
   final String studentId;
-  double? grade;
+  int? grade;
   String status;
 
   EnrolledCourseModel({
